Finish up sync
